update printing information
